pyproject.toml
[project]
name = "microbots"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "surrealdb>=1.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.1",
    "anthropic>=0.97.0",
    "composio>=0.11.6",
    "pydantic-ai[openai]>=1.87.0",
    "tiktoken>=0.12.0",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pyyaml>=6.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "e2e: end-to-end tests requiring a running SurrealDB",
    "live_llm: tests that make real LLM API calls",
]
testpaths = ["tests"]
